// src/interfaces/cbt.d.ts
export interface PrintCBTRow {
    rollNo: string;
    subCode: string;
    subName:string;
    branch: string;
    acYear: number;
    sem: number;
    regYear: string;
    branch: string;
    user:string;
}

export interface CBTSubjectsRow {
    subCode: string;
    subName:string;
    branch: string;
    acYear: number;
    sem: number;
    regYear: string;
}